Retrieval robustness: precise reranking, lossless splitting, OpenAI timeouts
- Reranker: exact-identifier queries score high only on explicit evidence,
  topical similarity is background, not proof (fixes trick-question lures)
- AbstractRagTool: optional averageDocumentTokens hint appended to the tool
  description so the LLM can size maxResults; cap raised to 50
- SnippetizerEnricher: prompt p3, long pages split into <=60k parts instead
  of truncation; shared executor with @PreDestroy; configurable OpenAI
  connect/read timeouts (enrichment.openai.*)
- Snippet ingesters: failed generation falls back to lossless plain-text
  chunks without generationKey stamp so the next update retries
- JavaApiCardRenderer.splitCard: line-aware splitting without dropping text
- UiSamplesIngester: HTTP timeouts on the crawler

Co-authored with a parallel ChatGPT review session.

+++ src/main/java/io/jmix/ai/backend/retrieval/SearchResultsFormatter.java
@@ -45,9 +45,9 @@ public class SearchResultsFormatter {
         if (tokens == null || tokens <= 0) {
             return documents;
         }
-        int budgetChars = tokens * CHARS_PER_TOKEN;
+        long budgetChars = (long) tokens * CHARS_PER_TOKEN;
         List<Document> result = new ArrayList<>();
-        int usedChars = 0;
+        long usedChars = 0;
         for (Document document : documents) {
             int length = document.getText() != null ? document.getText().length() : 0;
             if (!result.isEmpty() && usedChars + length > budgetChars) {
